# packages/api-contract/openapi.yaml
openapi: 3.1.0
info:
  title: CircuitLab API
  version: 1.0.0
  summary: Store digital logic circuits, simulate them, and page through their truth tables.
  description: |
    Conventions used everywhere in this API:

    - **Versioning.** Every path starts with `/v1`. Additive changes (new endpoints, new optional
      fields or parameters) stay in v1; a change that could break an existing client means `/v2`.
    - **Authentication.** Register at `/auth/register` or sign in at `/auth/login` to get an
      access token, and send it as `Authorization: Bearer <token>`. Access tokens expire after 15
      minutes; trade the refresh token for a new pair at `/auth/refresh` (each refresh token works
      once). Public circuits can be read without a token.
    - **Access.** A circuit is private to its owner until the owner shares it (as `viewer` or
      `editor`) or makes it public. A circuit you may not see answers 404, not 403, so private
      circuits can't be discovered by guessing ids. 403 means you can see it but may not do this.
    - **Errors.** Every error is an RFC 9457 problem document (`application/problem+json`) with a
      stable `code`. Validation errors list every problem at once in `issues`, each located by
      `pointer` (an RFC 6901 JSON Pointer into the JSON body), `parameter` (a query parameter),
      or `line`/`column` (a netlist).
    - **400 or 422.** 400 means the request could not be read: an unparseable body or a bad query
...
servers:
...

...
security:
  - bearerAuth: []
  - {}
...
paths:
  /circuits:
    get:
      operationId: listCircuits
      tags: [Circuits]
      summary: List circuits
      description: One page of circuit summaries. Fetch a single circuit to get its gates and wires.
      parameters:
        - $ref: "#/components/parameters/ListScope"
        - $ref: "#/components/parameters/PageLimit"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/CircuitSort"
        - $ref: "#/components/parameters/NameSearch"
      responses:
        "200":
          description: One page of circuits.
          headers:
            Link: { $ref: "#/components/headers/Link" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CircuitPage" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }

    post:
      operationId: createCircuit
      tags: [Circuits]
      summary: Create a circuit
      security:
        - bearerAuth: []
      description: |
        The new circuit belongs to you and is private. Send it as JSON, or as a netlist file (`text/vnd.circuitlab.netlist`, optionally
        gzip-compressed with `Content-Encoding: gzip`). A netlist's name is the `name` query
        parameter if given, else its `.name` line; `name` and `description` are only accepted with
        netlist bodies.

        Circuits with a feedback loop (such as a latch) are valid and are stored. They cannot be
        simulated combinationally; `summary.feedbackLoop` shows the loop.

        With `dryRun=true` nothing is stored: the answer is 200 with a validation report, or the
        same 422 a real request would get.
      parameters:
        - $ref: "#/components/parameters/DryRun"
        - $ref: "#/components/parameters/NetlistName"
        - $ref: "#/components/parameters/NetlistDescription"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CircuitInput" }
          text/vnd.circuitlab.netlist:
            schema: { $ref: "#/components/schemas/NetlistText" }
      responses:
        "201":
          description: Created.
          headers:
            Location: { $ref: "#/components/headers/Location" }
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Circuit" }
        "200":
          description: Dry run passed; nothing was stored.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ValidationReport" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "413": { $ref: "#/components/responses/ContentTooLarge" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "422": { $ref: "#/components/responses/InvalidContent" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
  /circuits/{circuitId}:
    parameters:
      - $ref: "#/components/parameters/CircuitId"
    get:
      operationId: getCircuit
      tags: [Circuits]
      summary: Get a circuit
      description: |
        Your circuits, circuits shared with you, and public circuits. JSON by default; send
        `Accept: text/vnd.circuitlab.netlist` to download it as a netlist.
      parameters:
        - $ref: "#/components/parameters/IfNoneMatch"
      responses:
        "200":
          description: The circuit.
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Circuit" }
            text/vnd.circuitlab.netlist:
              schema: { $ref: "#/components/schemas/NetlistText" }
        "304": { $ref: "#/components/responses/NotModified" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "406": { $ref: "#/components/responses/NotAcceptable" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }

    put:
      operationId: replaceCircuit
      tags: [Circuits]
      summary: Replace a circuit
      security:
        - bearerAuth: []
      description: |
        Replaces the name, description, gates, and wires; the version goes up by one. Accepts the same
        bodies as creating a circuit, including netlists. For the owner and editors.
      parameters:
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/NetlistName"
        - $ref: "#/components/parameters/NetlistDescription"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CircuitInput" }
          text/vnd.circuitlab.netlist:
            schema: { $ref: "#/components/schemas/NetlistText" }
      responses:
        "200":
          description: Replaced.
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Circuit" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "412": { $ref: "#/components/responses/PreconditionFailed" }
        "413": { $ref: "#/components/responses/ContentTooLarge" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "422": { $ref: "#/components/responses/InvalidContent" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }

    patch:
      operationId: updateCircuitMetadata
      tags: [Circuits]
      summary: Rename a circuit, change its description, or change who can see it
      security:
        - bearerAuth: []
      description: |
        A JSON Merge Patch (RFC 7386). Setting `description` to null removes it. The owner and
        editors may change the name and description; only the owner may change `visibility`.
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema: { $ref: "#/components/schemas/CircuitMetadataPatch" }
          application/json:
            schema: { $ref: "#/components/schemas/CircuitMetadataPatch" }
      responses:
        "200":
          description: Updated.
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Circuit" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "412": { $ref: "#/components/responses/PreconditionFailed" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "422": { $ref: "#/components/responses/InvalidContent" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }

    delete:
      operationId: deleteCircuit
      tags: [Circuits]
      summary: Delete a circuit
      description: For the owner only.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      responses:
        "204":
          description: Deleted.
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "412": { $ref: "#/components/responses/PreconditionFailed" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
...
  /circuits/{circuitId}/shares:
    parameters:
      - $ref: "#/components/parameters/CircuitId"
    get:
      operationId: listShares
      tags: [Sharing]
      summary: Who a circuit is shared with
      description: For the owner only.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Everyone the circuit is shared with, oldest share first.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ShareList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }

    post:
      operationId: shareCircuit
      tags: [Sharing]
      summary: Share a circuit, or change someone's role
      description: |
        For the owner only. The person needs an account; sharing again with the same person changes
        their role. A viewer may read and simulate the circuit; an editor may also change its name,
        description, gates, and wires. Only the owner may delete it, share it, or make it public.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ShareRequest" }
      responses:
        "201":
          description: Shared.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Share" }
        "200":
          description: Already shared with this person; their role is now the one given.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Share" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "422": { $ref: "#/components/responses/InvalidFields" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
  /circuits/{circuitId}/shares/{userId}:
    parameters:
      - $ref: "#/components/parameters/CircuitId"
      - $ref: "#/components/parameters/UserId"
    delete:
      operationId: unshareCircuit
      tags: [Sharing]
      summary: Stop sharing a circuit with someone
      description: The owner may remove anyone's access; anyone may remove their own.
      security:
        - bearerAuth: []
      responses:
        "204":
          description: They no longer have access through this share.
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: An access token from `/auth/login`, `/auth/register`, or `/auth/refresh`. Valid for 15 minutes.
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      description: An account's id, as in a share's `user.id`.
      schema:
        type: string
        minLength: 1
        maxLength: 64
    ListScope:
      name: scope
      in: query
      description: |
        Which circuits: `owned` (yours), `shared` (shared with you by others), or `public`
        (everyone's public circuits). Default: `owned` when signed in, `public` otherwise. `owned`
        and `shared` need a token.
      schema:
        type: string
        enum: [owned, shared, public]
    CircuitId:
      name: circuitId
      in: path
      required: true
      description: The circuit's id, as returned when it was created. Treat it as opaque.
      schema:
        type: string
        minLength: 1
        maxLength: 64
    PageLimit:
      name: limit
      in: query
      description: Circuits per page.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Cursor:
      name: cursor
      in: query
      description: Where to continue, taken from a previous page's `page.nextCursor`. Opaque.
      schema:
        title: a cursor copied from a previous page's page.nextCursor
        type: string
        maxLength: 512
        pattern: "^[A-Za-z0-9_-]+$"
    CircuitSort:
      name: sort
      in: query
      description: Order of the list. `-` means newest first. Ties are broken by id.
      schema:
        type: string
        enum: [-createdAt, -updatedAt, name]
        default: -createdAt
    NameSearch:
      name: q
      in: query
      description: Only circuits whose name contains this text, ignoring case.
      schema:
        type: string
        minLength: 1
        maxLength: 100
    DryRun:
      name: dryRun
      in: query
      description: Only validate; store nothing.
      schema:
        type: boolean
        default: false
    NetlistName:
      name: name
      in: query
      description: Name for a netlist body that has no `.name` line. Not accepted with JSON bodies.
      schema: { $ref: "#/components/schemas/CircuitName" }
    NetlistDescription:
      name: description
      in: query
      description: Description for a netlist body. Not accepted with JSON bodies.
      schema: { $ref: "#/components/schemas/CircuitDescription" }
...
    IfMatch:
      name: If-Match
      in: header
      description: The circuit's ETag as last seen. The request fails with 412 if it has changed since.
      schema:
        type: string
    IfNoneMatch:
      name: If-None-Match
      in: header
      description: An ETag already held. If it is still current, the answer is 304 with no body.
      schema:
        type: string
  headers:
    ETag:
      description: Identifies this version of the representation; it changes whenever the circuit does.
      schema:
        type: string
        examples: ['"7"']
    Location:
      description: URL of the new circuit.
      schema:
        type: string
        examples: [/v1/circuits/0199a3c4-5e6f-7a8b-9c0d-1e2f3a4b5c6d]
    Link:
      description: RFC 8288 links to neighbouring pages, e.g. `<...>; rel="next"`.
      schema:
        type: string
    RetryAfter:
      description: Seconds to wait before trying again.
      schema:
        type: integer
        minimum: 1
    WWWAuthenticate:
      description: How to authenticate (RFC 6750), e.g. `Bearer realm="circuitlab", error="invalid_token"`.
      schema:
        type: string
...
  responses:
    NotModified:
      description: The copy you hold (your `If-None-Match` ETag) is still current.
    InvalidRequest:
      description: The request could not be read (`invalid-request`, `malformed-body`).
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Unauthorized:
      description: |
        Sign in first (`unauthenticated`), the token can't be used (`invalid-token`: refresh it
        or sign in again), or a sign-in failed (`invalid-credentials`).
      headers:
        WWW-Authenticate: { $ref: "#/components/headers/WWWAuthenticate" }
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Forbidden:
      description: You can see this circuit but may not do this to it (`forbidden`).
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    InvalidFields:
      description: The body was read, but a field is invalid (`invalid-fields`); see `issues`.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    NotFound:
      description: |
        No circuit has this id, or you may not see it (`not-found`). The two are deliberately
        indistinguishable.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotAcceptable:
      description: None of the formats in `Accept` is available here (`not-acceptable`).
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    PreconditionFailed:
      description: The circuit changed since you last read it (`precondition-failed`).
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    ContentTooLarge:
      description: The body is over 5 MB (`content-too-large`).
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    UnsupportedMediaType:
      description: The body's `Content-Type` is not accepted here (`unsupported-media-type`).
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    InvalidContent:
      description: |
        The request was read, but what it describes is invalid: `invalid-circuit`, `invalid-netlist`,
        `invalid-inputs`, `feedback-loop`, `too-many-inputs`, or `computation-too-large`.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    ServiceUnavailable:
      description: |
        Try again later: `server-busy` (all simulation workers busy), `server-unavailable` (the
        server is shutting down, or the database can't be reached), or `simulation-timeout`. See
        `Retry-After`. Any operation can answer this: every one of them uses the database.
      headers:
        Retry-After: { $ref: "#/components/headers/RetryAfter" }
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
...
    Gate:
      description: |
        One gate, told apart by `type`, exactly like the engine's TypeScript union. INPUT and CONST
        have no input pins; OUTPUT, BUF, and NOT have one; the others have 2 to 64. Only CONST
        gates have a `value`.
      type: object
      oneOf:
        - $ref: "#/components/schemas/InputGate"
        - $ref: "#/components/schemas/OutputGate"
        - $ref: "#/components/schemas/ConstGate"
        - $ref: "#/components/schemas/LogicGate"
      discriminator:
        propertyName: type
        mapping:
          INPUT: "#/components/schemas/InputGate"
          OUTPUT: "#/components/schemas/OutputGate"
          CONST: "#/components/schemas/ConstGate"
          BUF: "#/components/schemas/LogicGate"
          NOT: "#/components/schemas/LogicGate"
          AND: "#/components/schemas/LogicGate"
          OR: "#/components/schemas/LogicGate"
          NAND: "#/components/schemas/LogicGate"
          NOR: "#/components/schemas/LogicGate"
          XOR: "#/components/schemas/LogicGate"
          XNOR: "#/components/schemas/LogicGate"
...
    Wire:
      type: object
      description: Connects the output of gate `from` to input pin `toPin` (numbered from 0) of gate `to`.
      required: [from, to, toPin]
      additionalProperties: false
      properties:
        id: { $ref: "#/components/schemas/GateId" }
        from: { $ref: "#/components/schemas/GateId" }
        to: { $ref: "#/components/schemas/GateId" }
        toPin:
          type: integer
          minimum: 0
          maximum: 63
    CircuitName:
      title: a name that is not blank, at most 200 characters
      type: string
      minLength: 1
      maxLength: 200
      pattern: "\\S"
    CircuitDescription:
      type: string
      maxLength: 2000
    NetlistText:
      type: string
      description: A CircuitLab netlist, one gate per line, e.g. `sum = XOR(A, B)`.
      examples:
        - ".name \"Half adder\"\nA = INPUT\nB = INPUT\nsum = XOR(A, B)\ncarry = AND(A, B)\nS = OUTPUT(sum)\nC = OUTPUT(carry)\n"
    CircuitInput:
      type: object
      description: A circuit as sent by a client.
      required: [name, gates, wires]
      additionalProperties: false
      properties:
        name: { $ref: "#/components/schemas/CircuitName" }
        description: { $ref: "#/components/schemas/CircuitDescription" }
        gates:
          type: array
          maxItems: 10000
          items: { $ref: "#/components/schemas/Gate" }
        wires:
          type: array
          maxItems: 50000
          items: { $ref: "#/components/schemas/Wire" }
      examples:
        - name: Half adder
          gates:
            - { id: A, type: INPUT }
            - { id: B, type: INPUT }
            - { id: sum, type: XOR }
            - { id: carry, type: AND }
            - { id: S, type: OUTPUT, label: Sum }
            - { id: C, type: OUTPUT, label: Carry }
          wires:
            - { from: A, to: sum, toPin: 0 }
            - { from: B, to: sum, toPin: 1 }
            - { from: A, to: carry, toPin: 0 }
            - { from: B, to: carry, toPin: 1 }
            - { from: sum, to: S, toPin: 0 }
            - { from: carry, to: C, toPin: 0 }
    CircuitMetadataPatch:
      type: object
      minProperties: 1
      additionalProperties: false
      properties:
        name: { $ref: "#/components/schemas/CircuitName" }
        description:
          description: A new description, or null to remove it.
          type: [string, "null"]
          maxLength: 2000
        visibility: { $ref: "#/components/schemas/Visibility" }
    Visibility:
      description: |
        Who may see the circuit besides its owner and the people it is shared with: nobody
        (`private`, the default) or everyone (`public`, read-only).
      type: string
      enum: [private, public]
    UserSummary:
      type: object
      required: [id, displayName]
      properties:
        id: { type: string }
        displayName: { $ref: "#/components/schemas/DisplayName" }
    CircuitSummary:
      type: object
      required: [gates, wires, inputs, outputs, feedbackLoop]
      properties:
        gates: { type: integer, minimum: 0 }
        wires: { type: integer, minimum: 0 }
        inputs:
          description: INPUT gate ids, in declaration order; the keys a simulation needs.
          type: array
          items: { $ref: "#/components/schemas/GateId" }
        outputs:
          description: OUTPUT gate ids, in declaration order.
          type: array
          items: { $ref: "#/components/schemas/GateId" }
        feedbackLoop:
          description: A loop of gates (first id repeated at the end), or null if there is none.
          type: [array, "null"]
          items: { $ref: "#/components/schemas/GateId" }
    CircuitListItem:
      type: object
      required: [id, name, owner, visibility, version, createdAt, updatedAt, summary]
      properties:
        id: { type: string }
        name: { $ref: "#/components/schemas/CircuitName" }
        description: { $ref: "#/components/schemas/CircuitDescription" }
        owner: { $ref: "#/components/schemas/UserSummary" }
        visibility: { $ref: "#/components/schemas/Visibility" }
        version:
          description: Starts at 1 and goes up by one with every change.
          type: integer
          minimum: 1
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        summary: { $ref: "#/components/schemas/CircuitSummary" }
    Circuit:
      allOf:
        - { $ref: "#/components/schemas/CircuitListItem" }
        - type: object
          required: [gates, wires]
          properties:
            gates:
              type: array
              items: { $ref: "#/components/schemas/Gate" }
            wires:
              type: array
              items: { $ref: "#/components/schemas/Wire" }
    CircuitPage:
      type: object
      required: [items, page, links]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/CircuitListItem" }
        page:
          type: object
          required: [limit, nextCursor]
          properties:
            limit: { type: integer }
            nextCursor:
              description: Pass as `cursor` to get the next page; null on the last page.
              type: [string, "null"]
        links:
          type: object
          required: [self, next]
          properties:
            self: { type: string }
            next: { type: [string, "null"] }
    ValidationReport:
      type: object
      required: [valid, summary]
      properties:
        valid: { const: true }
        summary: { $ref: "#/components/schemas/CircuitSummary" }
...
    ShareRole:
      description: "`viewer` may read and simulate; `editor` may also change the circuit."
      type: string
      enum: [viewer, editor]
    ShareRequest:
      type: object
      required: [email, role]
      additionalProperties: false
      properties:
        email:
          description: The email address of the person's account.
          type: string
          minLength: 1
          maxLength: 254
        role: { $ref: "#/components/schemas/ShareRole" }
      examples:
        - { email: bob@example.com, role: viewer }
    Share:
      type: object
      required: [user, role, createdAt]
      properties:
        user:
          allOf:
            - { $ref: "#/components/schemas/UserSummary" }
            - type: object
              required: [email]
              properties:
                email: { type: string }
        role: { $ref: "#/components/schemas/ShareRole" }
        createdAt:
          description: When it was first shared with this person.
          type: string
          format: date-time
    ShareList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Share" }
    ProblemIssue:
      type: object
      description: One problem found while validating. Exactly one of the location fields is usually set.
      required: [code, message]
      properties:
        code:
          description: Machine-readable, e.g. `UNKNOWN_FIELD`, `UNCONNECTED_PIN`, `MISSING_INPUT`.
          type: string
        message: { type: string }
        pointer:
          description: RFC 6901 JSON Pointer into the request body, e.g. `/gates/3/type`.
          type: string
        parameter:
          description: The query parameter at fault.
          type: string
        line: { type: integer, minimum: 1 }
        column: { type: integer, minimum: 1 }
    Problem:
      type: object
      description: An RFC 9457 problem document.
      required: [type, title, status, code]
      properties:
        type:
          description: Identifies the kind of problem; a relative URL such as `/problems/invalid-circuit`.
          type: string
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance:
          description: The request path that failed.
          type: string
        code:
          description: The last segment of `type`, for easy matching in code.
          type: string
          enum:
            - invalid-request
            - malformed-body
            - unauthenticated
            - invalid-token
            - invalid-credentials
            - forbidden
            - not-found
            - not-acceptable
            - email-taken
            - version-conflict
            - precondition-failed
            - content-too-large
            - unsupported-media-type
            - invalid-circuit
            - invalid-netlist
            - invalid-inputs
            - feedback-loop
            - too-many-inputs
            - computation-too-large
            - invalid-fields
            - too-many-requests
            - internal-error
            - server-busy
            - server-unavailable
            - simulation-timeout
        issues:
          type: array
          items: { $ref: "#/components/schemas/ProblemIssue" }
        cycle:
          description: For `feedback-loop`, the gates around the loop.
          type: array
          items: { $ref: "#/components/schemas/GateId" }